Refactor genetic-algorithm lib to individual files
